tests: refactor go tests to not depend on each other
